Address OpenHENS PR review comments
