+++ OpenPinch/lib/__init__.py
@@ -22,7 +22,9 @@ from .enums import (
     CogenerationTarget,
     FluidPhase,
     GraphType,
+    HeatExchangerKind,
     HeatExchangerNetworkLabel,
+    HeatExchangerStreamRole,
     HeatExchangerTypes,
     HeatFlowUnits,
     HeatPump,
@@ -139,7 +141,9 @@ __all__ = [
     "GraphType",
     "FluidPhase",
     "GT",
+    "HeatExchangerKind",
     "HeatExchangerNetworkLabel",
+    "HeatExchangerStreamRole",
     "HEN",
     "HeatExchangerTypes",
     "HeatFlowUnits",
